Initial commit
Started project with a working prototype, but with only the
most basic features. It displays correctly the response of
systems with no, low and high damping, and also provides a
reset function, all in an ugly green-on-black screen.

I understand the missing features, but this is enough for
my presentation.