Add marks to show previous positions
Added small blue circles that are spawned in the middle of
the rectangle-mass-thingy and are moved to the right at
every time the mass is moved (preciselly, at every tick of
the update-time timer). The spawn time is marked by a
separeted timer, spawn-dots, as using the main didn't work.
Also, this is good to adjust with necessity for a
better-looking line for the graphic.